Overhaul tests
Handle new instance list JSON structure

Serialized instance list is now a JSON array of [host, {instance-data}]

Replace stdlib spec lib with Spectator

match_array allows comparing two arrays with its internal objects
being in any order

Failures on match array will also show where the two arrays fails to
match up

Use service setter to specify mock instance data

Prior to this commit we used ADI.bind to inject the location of the
mock file into MockQueryInstance. However, this means that the
mock file is now set for the entire application instead of an
individual test, and subsequent bind calls to replace it will be
ignored.

This commit replaces the usage of ADI.bind with a setter on the mock
service instead allowing the mock file to be customized for each test